# iruyan-api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a server for IRUYAN.",
        "title": "IRUYAN API",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/worktime/entry": {
            "post": {
                "description": "Creates a work time record when a user enters a room",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "worktime"
                ],
                "summary": "Record entry time",
                "parameters": [
                    {
                        "type": "string",
                        "default": "johndoe",
                        "description": "Iruyan ID",
                        "name": "iruyanId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Room UUID",
                        "name": "roomId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Task description",
                        "name": "task",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.WorkTimeResponseSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/worktime/fetch/all": {
            "get": {
                "description": "Retrieves all work time records from the database",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "worktime"
                ],
                "summary": "Get all WorkTime records",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.WorkTimeListResponseSwagger"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/worktime/fetch/{iruyanId}": {
            "get": {
                "description": "Retrieves all work time records for a user identified by iruyanID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "worktime"
                ],
                "summary": "Get WorkTime records for a specific user",
                "parameters": [
                    {
                        "type": "string",
                        "default": "johndoe",
                        "description": "Iruyan ID",
                        "name": "iruyanId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.WorkTimeListResponseSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/worktime/latest": {
            "get": {
                "description": "Retrieves the latest work time record for a user that has not ended (leaving_time is null)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "worktime"
                ],
                "summary": "Get latest active entry",
                "parameters": [
                    {
                        "type": "string",
                        "default": "johndoe",
                        "description": "Iruyan ID",
                        "name": "iruyanId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Room UUID",
                        "name": "roomId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/responses.WorkTimeResponseSwagger"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/worktime/recent": {
            "get": {
                "description": "Retrieves the most recent N work time records for a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "worktime"
                ],
                "summary": "Get recent work logs",
                "parameters": [
                    {
                        "type": "string",
                        "default": "johndoe",
                        "description": "Iruyan ID",
                        "name": "iruyanId",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 5,
                        "description": "Number of records to return",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.WorkTimeResponseSwagger"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/worktime/weekly": {
            "get": {
                "description": "Retrieves all work time records from the past 7 days for a user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "worktime"
                ],
                "summary": "Get weekly work logs",
                "parameters": [
                    {
                        "type": "string",
                        "default": "johndoe",
                        "description": "Iruyan ID",
                        "name": "iruyanId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/responses.WorkTimeResponseSwagger"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/responses.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "responses.WorkTimeListResponseSwagger": {
            "type": "object",
            "properties": {
                "work_times": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/responses.WorkTimeResponseSwagger"
                    }
                }
            }
        },
...
        "responses.WorkTimeResponseSwagger": {
            "type": "object",
            "properties": {
                "duration": {
                    "description": "formatted like \"1h23m\"",
                    "type": "string"
                },
                "entryTime": {
                    "description": "RFC3339",
                    "type": "string"
                },
                "iruyanId": {
                    "type": "string"
                },
                "leavingTime": {
                    "description": "RFC3339 or nil",
                    "type": "string"
                },
                "roomId": {
                    "type": "string"
                },
                "task": {
                    "type": "string"
                }
            }
        }
    }
}